requirements.txt
adafruit-io==2.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Hello-World-Package==0.1.3
idna==3.7
paho-mqtt==2.1.0
python-dotenv==1.0.1
requests==2.32.2
setuptools==70.0.0
urllib3==2.2.1
